requirements.txt
asgiref==3.10.0
coverage==7.11.0
Django==5.2.7
django-cors-headers==4.9.0
django-extensions==3.2.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
iniconfig==2.3.0
markdown-it-py==4.0.0
mdurl==0.1.2
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
pytest-sugar==1.1.1
python-dotenv==1.1.1
rich==14.2.0
sqlparse==0.5.3
termcolor==3.1.0
gunicorn==21.2.0
dj-database-url==2.1.0
whitenoise==6.6.0